Refactor: Co-locate feature-specific components
- Moved apartment-filters.tsx into app/apartments/components/
- Moved all components from components/settings/ into app/settings/components/
- Updated all relevant import paths.

This change improves modularity and makes it easier to navigate the codebase by grouping components with the features they belong to.
