+++ app/settings/components/security-settings.tsx
@@ -480,4 +480,3 @@ export default function SecuritySettings({ onSave }: SecuritySettingsProps) {
     </div>
   )
 }
-
